Move test files to test directory
